backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
mnemonic
SpeechRecognition
pydub
google-generativeai

